+++ lib/cocoapods/installer/installation_options.rb
@@ -106,6 +106,7 @@ module Pod
       option :deterministic_uuids, true
       option :integrate_targets, true
       option :lock_pod_sources, true
+      option :warn_for_multiple_pod_sources, true
       option :share_schemes_for_development_pods, false
 
       module Mixin