+++ lib/cocoapods/resolver.rb
@@ -2,10 +2,20 @@ require 'molinillo'
 require 'cocoapods/resolver/lazy_specification'
 
 module Pod
+  class NoSpecFoundError < Informative
+    def exit_status
+      @exit_status ||= 31
+    end
+  end
+
   # The resolver is responsible of generating a list of specifications grouped
   # by target for a given Podfile.
   #
   class Resolver
+    include Pod::Installer::InstallationOptions::Mixin
+
+    delegate_installation_options { podfile }
+
     # @return [Sandbox] the Sandbox used by the resolver to find external
     #         dependencies.
     #
@@ -288,7 +298,7 @@ module Pod
     def specifications_for_dependency(dependency, additional_requirements = [])
       requirement = Requirement.new(dependency.requirement.as_list + additional_requirements)
       find_cached_set(dependency).
-        all_specifications.
+        all_specifications(installation_options.warn_for_multiple_pod_sources).
         select { |s| requirement.satisfied_by? s.version }.
         map { |s| s.subspec_by_name(dependency.name, false) }.
         compact.
@@ -385,6 +395,7 @@ module Pod
     #
     def handle_resolver_error(error)
       message = error.message.dup
+      type = Informative
       case error
       when Molinillo::VersionConflict
         error.conflicts.each do |name, conflict|
@@ -423,6 +434,7 @@ module Pod
             found_conflicted_specs = conflicts.reject { |c| search_for(c).empty? }
             if found_conflicted_specs.empty?
               # There are no existing specification inside any of the spec repos with given requirements.
+              type = NoSpecFoundError
               dependencies = conflicts.count == 1 ? 'dependency' : 'dependencies'
               message << "\n\nNone of your spec sources contain a spec satisfying "\
                 "the #{dependencies}: `#{conflicts.join(', ')}`." \
@@ -439,7 +451,7 @@ module Pod
           end
         end
       end
-      raise Informative, message
+      raise type, message
     end
 
     # Returns whether the given spec is platform-compatible with the dependency